chore(core): remove @bara/core version source code
