--- packages/barn/.clang-format
@@ -1,3 +1,0 @@
-Language:        JavaScript
-BasedOnStyle:    Google
-ColumnLimit:     80
