--- packages/barn/CHANGELOG.md
@@ -1,77 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [2.5.0](https://github.com/barajs/bara/compare/v2.4.0...v2.5.0) (2019-08-10)
-
-
-### Bug Fixes
-
-* **deps:** update dependency dot-prop to v5.0.1 ([8f2ded5](https://github.com/barajs/bara/commit/8f2ded5))
-* **deps:** update dependency dot-prop to v5.1.0 ([81dfb77](https://github.com/barajs/bara/commit/81dfb77))
-
-
-### Features
-
-* **barn:** `useBarn` now return value at the time it is called ([be14a6e](https://github.com/barajs/bara/commit/be14a6e)), closes [#56](https://github.com/barajs/bara/issues/56)
-
-
-
-
-
-# [2.4.0](https://github.com/barajs/bara/compare/v2.3.2...v2.4.0) (2019-04-24)
-
-
-### Bug Fixes
-
-* **barn:** not emit init state and latest value ([00aa16d](https://github.com/barajs/bara/commit/00aa16d)), closes [#47](https://github.com/barajs/bara/issues/47)
-
-
-
-
-
-## [2.3.2](https://github.com/barajs/bara/compare/v2.3.1...v2.3.2) (2019-04-22)
-
-**Note:** Version bump only for package @bara/barn
-
-
-
-
-
-## [2.3.1](https://github.com/barajs/bara/compare/v2.3.0...v2.3.1) (2019-04-19)
-
-**Note:** Version bump only for package @bara/barn
-
-
-
-
-
-# [2.3.0](https://github.com/barajs/bara/compare/v2.2.0...v2.3.0) (2019-04-19)
-
-
-### Bug Fixes
-
-* **basics:** remove useInitStream from hook ([5040311](https://github.com/barajs/bara/commit/5040311))
-
-
-### Features
-
-* **barn:** new operator `setBarnState` based on `useEmitter` ([d4a2bd0](https://github.com/barajs/bara/commit/d4a2bd0)), closes [#35](https://github.com/barajs/bara/issues/35)
-
-
-
-
-
-# [2.2.0](https://github.com/barajs/bara/compare/v2.1.0...v2.2.0) (2019-04-10)
-
-
-### Bug Fixes
-
-* **barn:** replace `fbemitter` with `xstream` because of license ([b5cb6e0](https://github.com/barajs/bara/commit/b5cb6e0))
-
-
-### Features
-
-* **barn:** able to set state and handle key change dynamically ([b3faace](https://github.com/barajs/bara/commit/b3faace)), closes [#14](https://github.com/barajs/bara/issues/14)
-* **barn:** able to setState and useBarn by property path ([8fc2ecc](https://github.com/barajs/bara/commit/8fc2ecc)), closes [#14](https://github.com/barajs/bara/issues/14)
