--- packages/barn/README.md
@@ -1,11 +1,0 @@
-# Bara Barn
-
-Basic state management of a Bara application.
-
-## Installation
-
-```
-yarn add @bara/barn
-
-npm i --save @bara/barn
-```
