--- packages/barn/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "@bara/barn",
-  "version": "2.5.0",
-  "description": "BaraJS Internal State Management Module",
-  "license": "MIT",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "keywords": [],
-  "files": [
-    "dist"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "watch": {
-    "lint": "{src}/*.ts"
-  },
-  "scripts": {
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "compile": "tsc -p ./tsconfig.build.json",
-    "compile:watch": "tsc -w -p ./tsconfig.build.json",
-    "lint": "tslint -p .",
-    "format": "prettier --write '{.,src/**}/*.{js,jsx,ts,tsx}'",
-    "prepare": "yarn compile",
-    "pretest": "yarn format",
-    "lint:watch": "npm-watch",
-    "dev": "concurrently --kill-others \"yarn test:watch\" \"yarn lint:watch\" \"yarn compile:watch\" "
-  },
-  "devDependencies": {
-    "@types/dot-prop": "4.2.0",
-    "@types/fbemitter": "2.0.32",
-    "@types/jest": "24.0.18",
-    "concurrently": "4.1.2",
-    "jest": "24.9.0",
-    "npm-watch": "0.6.0",
-    "prettier": "1.18.2",
-    "ts-jest": "24.1.0",
-    "tslint": "5.20.0",
-    "tslint-config-airbnb": "5.11.1",
-    "tslint-config-prettier": "1.18.0",
-    "tslint-react": "4.0.0",
-    "typescript": "3.6.3"
-  },
-  "dependencies": {
-    "@bara/basics": "^2.5.0",
-    "@bara/core": "^2.5.0",
-    "dot-prop": "5.1.0"
-  },
-  "gitHead": "f4c8920813b598869d4e664195ca7838532dd17f"
-}
