--- packages/barn/tsconfig.build.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "target": "es2015",
-    "module": "commonjs",
-    "allowJs": false,
-    "composite": true,
-    "declaration": true,
-    "noEmit": false,
-    "outDir": "dist/",
-    "rootDir": "src"
-  },
-  "include": ["src/*.ts", "src/**/*.ts"]
-}
