+++ packages/core/src/chain/cond.ts
