--- packages/core/src/model/app.ts
@@ -1,4 +1,0 @@
-import { Stream } from 'xstream'
-import { BaraStreamPayload } from './stream'
-
-export type AppStream<T> = Stream<BaraStreamPayload<T>>
