--- packages/core/src/model/base.ts
@@ -1,5 +1,0 @@
-export interface Base {
-  name: string | null
-}
-
-export type Use<T> = (...args: any[]) => T
