--- packages/core/src/model/index.ts
@@ -1,8 +1,0 @@
-export * from './action'
-export * from './app'
-export * from './base'
-export * from './condition'
-export * from './event'
-export * from './emitter'
-export * from './stream'
-export * from './trigger'
