+++ packages/core/src/portion.ts
