+++ packages/core/src/stream.ts
