+++ packages/core/src/type.ts
