--- packages/core/src/xstream.ts
@@ -1,5 +1,0 @@
-import xs from 'xstream'
-
-export { xs }
-
-export * from 'xstream'
